Move format-srcset to /src/utils/
