+++ src/utils/format-srcset/index.test.js
